requirements.txt
Django==5.1.5
pillow==9.5.0
requests==2.31.0
dj-database-url==2.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular